MAINT:lib: group classes in different packages
The current layout of the classes was rather "flat", this resulted in
a class hierarchy that was hard to follow from the filesystem structure.
Separate in-toto into a models package, and the key-related code into a
keys package.

For the time being I'll leave App.java in .lib, as we may later use that
package for the "exposed" api.