ENH:keys: add Abstract superclass
The current RSAKey implementation is rather opinionated. Further, in the
forseeable future we may need to add support for other keytypes (e.g.,
DSA). As such, a templating superclass will help standarize the type
interfaces with the rest of the application. Add an abstract Key
superclass and define a any missing methods in RSAKey.