ENH:models.Metablock: add dump method
In-toto metadata needs to be serialized somehow, with the easiest
transport being filesystem storage. Add a method to dump the link
metadata to the filesystem using an arbitrary name.

Althogh the specification requires to use a specific name (derived from
the signing keyid and the role that's being signed), we won't set that
restriction for now. We will add the logic to compute the filename
automatically on a later patch.