ENH:lib.JSONEncoder: add canonicalization methods
In order to hash (and therefore sign) in-toto link metadata, it's
necessary to encode the metdata into canonical JSON. This requires a
little bit more effort than regular JSON, and this computation is
required in multiple places (e.g., keyid computation and signable
signatures), so a centralized location to perform this operation is
recommended.

Generate a interface with a default encoding method that other classes
can leverage to compute the canonical representation of themselves.