ENH:TestRSAKey: add keyid computation tests
Ensuring the keyid is correctly computed is paramount to allow for
interoperability. Even though any deterministic computation would be
enough for the Java implementation, it may be possible that the same
KeyId's are not computed properly by other implementations and thus
break verification on either side.

Add a test to ensure the keyid for a particular RSA key produced using
the Python version of In-toto yields the same KeyID.