FIX:models: Link metadata wasn't in-toto compliant
I have to embarassingly admit that a couple of typos and assumptions I
made yielded the previous implementation of link metadata not in-toto
compliant. This meant that it couldn't be de-serialized appropriately in
the python implementation.

Add all the missing fields, rename the wrong ones and change the data
types of the rest. Specificaly:

    + byproducts field
    - materials/products -> are a dictionary now
    - command -> is an array now
    - env -> renamed to environment, changed to dictionary