ENH:link: refactor class hierarcy
The api required users to instantiate a "Metablock" class to perform
signing in the metadata. This was rather counter intuitive and hard to
explain. Fix the class hierarcy so that Link inherits from metablock,
and contains a LinkSignable class within it.

Likewise, update the access level for the rest of the model, so that
only Links and Artifacts can be instantiated and inherited from.

Finally, extend the api to simplify class accessors and update the
readme to reflect these changes..