FIX:JSONENcoder: Disable HTML escaping on Gson
The default GSON behavior is to escape certain HTML-relevant characters
(e.g. < or =) so as to avoid it from causing parsing errors. However,
this created issues in canonical-json representation given that the
keyid computed would not match the one in the python implementation.
Disable HTML escaping in the GSONBuilder call to avoid uneccessary
escaping.