Add json serialize adpter for metablock dump
Override gson serializer to convert all numbers to non-floating
point numbers, as we don't allow floating point numbers in the
reference implementation.  This is required to handle generic data,
such as `byproducts` or `environment`, where we don't declare the
type of the contained values. Gson treats any numeric value where
the type is not specified in the target data structure as double.